refactor: standardize directory naming to kebab-case format
- Rename PascalCase directories to kebab-case for consistency
- Update performance-api example directories:
  - ActualFmpEntry -> actual-fmp-entry
  - HostPlatformTiming -> host-platform-timing
  - InitBackgroundRuntimeEntry -> init-background-runtime-entry
  - InitContainerEntry -> init-container-entry
  - InitLynxViewEntry -> init-lynxview-entry
  - LazyBundleEntry -> lazy-bundle-entry
  - LoadBundleEntry -> load-bundle-entry
  - PerformanceObserverDisconnect -> performance-observer-disconnect
  - PerformanceObserverObserve -> performance-observer-observe
  - PipelineEntry -> pipeline-entry
  - ReloadBundleEntry -> reload-bundle-entry
  - SimpleObserve -> simple-observe
- Maintain file functionality while improving code organization
