+++ examples/performance-api/lynx.config.mjs
@@ -11,20 +11,20 @@ import { pluginTypeCheck } from "@rsbuild/plugin-type-check";
 export default defineConfig({
   source: {
     entry: {
-      FcpEntry: "./src/FcpEntry/index.tsx",
-      ActualFmpEntry: "./src/ActualFmpEntry/index.tsx",
-      PipelineEntry: "./src/PipelineEntry/index.tsx",
-      HostPlatformTiming: "./src/HostPlatformTiming/index.tsx",
-      ReloadBundleEntry: "./src/ReloadBundleEntry/index.tsx",
-      LazyBundleEntry: "./src/LazyBundleEntry/index.tsx",
-      LoadBundleEntry: "./src/LoadBundleEntry/index.tsx",
-      InitLynxviewEntry: "./src/InitLynxviewEntry/index.tsx",
-      InitBackgroundRuntimeEntry: "./src/InitBackgroundRuntimeEntry/index.tsx",
-      InitContainerEntry: "./src/InitContainerEntry/index.tsx",
-      CreateCustomMetric: "./src/CreateCustomMetric/index.tsx",
-      PerformanceObserverObserve: "./src/PerformanceObserverObserve/index.tsx",
-      PerformanceObserverDisconnect: "./src/PerformanceObserverDisconnect/index.tsx",
-      SimpleObserve: "./src/SimpleObserve/index.tsx",
+      "fcp-entry": "./src/fcp-entry/index.tsx",
+      "actual-fmp-entry": "./src/actual-fmp-entry/index.tsx",
+      "pipeline-entry": "./src/pipeline-entry/index.tsx",
+      "host-platform-timing": "./src/host-platform-timing/index.tsx",
+      "reload-bundle-entry": "./src/reload-bundle-entry/index.tsx",
+      "lazy-bundle-entry": "./src/lazy-bundle-entry/index.tsx",
+      "load-bundle-entry": "./src/load-bundle-entry/index.tsx",
+      "init-lynx-view-entry": "./src/init-lynx-view-entry/index.tsx",
+      "init-background-runtime-entry": "./src/init-background-runtime-entry/index.tsx",
+      "init-container-entry": "./src/init-container-entry/index.tsx",
+      "create-custom-metric": "./src/create-custom-metric/index.tsx",
+      "performance-observer-observe": "./src/performance-observer-observe/index.tsx",
+      "performance-observer-disconnect": "./src/performance-observer-disconnect/index.tsx",
+      "simple-observe": "./src/simple-observe/index.tsx",
     },
   },
   plugins: [
