+++ examples/performance-api/src/common/scroll-item/index.scss
