+++ examples/performance-api/src/fcp-entry/index.scss
