+++ examples/performance-api/src/host-platform-timing/index.scss
