+++ examples/performance-api/src/init-background-runtime-entry/index.tsx
