+++ examples/performance-api/src/init-container-entry/index.tsx
