+++ examples/performance-api/src/init-lynx-view-entry/index.css
