+++ examples/performance-api/src/lazy-bundle-entry/MyLazyBundle.css
