+++ examples/performance-api/src/lazy-bundle-entry/MyLazyBundle.tsx
