+++ examples/performance-api/src/lazy-bundle-entry/index.scss
