+++ examples/performance-api/src/load-bundle-entry/index.scss
