+++ examples/performance-api/src/performance-observer-disconnect/index.tsx
