+++ examples/performance-api/src/performance-observer-observe/index.tsx
