+++ examples/performance-api/src/pipeline-entry/index.scss
