+++ examples/performance-api/src/reload-bundle-entry/index.scss
