+++ examples/performance-api/src/simple-observe/index.tsx
